requirements.txt
django
gunicorn
whitenoise
psycopg2-binary
requests
urllib3
pymysql
djangorestframework
django-cors-headers
channels
channels-redis
daphne